Rename index.js -> main.js now that less is done there
