+++ README.md
@@ -148,11 +148,11 @@ You don't _need_ to set up all of these services, but if you don't, their associ
 # Under the Hood
 At the highest level, the bot listens to a stream of messages, calling the `handleMessage` function when one is received. This function has two main tasks: (1) parse the message to determine which (more specific) handler function it should be passed to and (2) update the information associated with the group in memory. These tasks are performed in parallel, and if no information is currently stored about the thread, it is initialized in the database. The database also stores the appstate after logging in so that a hard user/password login doesn't have to be performed every time. To purge this appstate, use `make logout` or run [`logout.js`](logout.js).
 
-There are three main types of messages to handle: pings, Easter eggs, and commands. All of the associated handling functions (`handlePings`, `handleEasterEggs`, and `handleCommand`) are available externally by requiring the index module. If a message contains a ping, the named member(s) will be notified in a private message thread with the bot. Easter eggs are a set of hidden responses from the bot that can be configured in [`easter.js`](easter.js). These are off by default. Commands are the main feature of the bot and comprise the majority of its codebase.
+There are three main types of messages to handle: pings, Easter eggs, and commands. All of the associated handling functions (`handlePings`, `handleEasterEggs`, and `handleCommand`) are available externally by requiring the main module. If a message contains a ping, the named member(s) will be notified in a private message thread with the bot. Easter eggs are a set of hidden responses from the bot that can be configured in [`easter.js`](easter.js). These are off by default. Commands are the main feature of the bot and comprise the majority of its codebase.
 
 The bot's command structure is "context-free"; it doesn't care where in the message the trigger word is used and what comes before it – as a result, only the text following the trigger word is passed to the `handleCommand`. The user ID of the sender, the `groupInfo` object for the thread, and the full message object from the listener are also passed.
 
-The `groupInfo` object is a record of the information stored in the database for a given thread, and it is passed to most utility functions used in [`index.js`](index.js) by `handleCommand`. Its structure changes with the internals of Facebook's message representation and the facebook-chat-api's parsing of it, but it is currently represented as follows:
+The `groupInfo` object is a record of the information stored in the database for a given thread, and it is passed to most utility functions used in [`main.js`](main.js) by `handleCommand`. Its structure changes with the internals of Facebook's message representation and the facebook-chat-api's parsing of it, but it is currently represented as follows:
 
 ```js
 let groupInfo = {
