+++ package.json
@@ -2,12 +2,12 @@
   "name": "assumezero-bot",
   "version": "2.0.0",
   "description": "A group chat bot that performs a variety of functions.",
-  "main": "index.js",
+  "main": "main.js",
   "engines": {
     "node": "8.0.0"
   },
   "scripts": {
-    "start": "node index.js",
+    "start": "node main.js",
     "test": "node tests/tests.js"
   },
   "author": "Cameron Bernhardt (AstroCB)",
