+++ server.js
@@ -2,7 +2,7 @@ const express = require("express");
 const http = require("http");
 const app = express();
 const bodyParser = require("body-parser");
-const main = require("./index");
+const main = require("./main");
 const config = require("./config");
 
 app.set("port", (process.env.PORT || 3000));
